// next/modules/core/api/model/guide/guideDetailedData.ts
/**
 * Go2kavkaz Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: hello@icerockdev.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ScheduleDate} from '../common/scheduleDate';
import {TravelMode} from '../common/travelMode';

export interface GuideDetailedData {
  id: number;
  age: number;
  /**
   * Guide avatar image key
   */
  avatar: {key: string; url: string};
  /**
   * Array minio keys
   */
  photos: Array<{key: string; url: string}>;
  /**
   * Guide Name
   */
  name: string;
  /**
   * About of guide
   */
  description: string;
  /**
   * Max participants count
   */
  maxNumberOfParticipants: number;
  /**
   * Is it possible with children or not
   */
  withChildren?: boolean;
  /**
   * Maximum duration of work
   */
  duration: number;
  /**
   * Cost per 1 hour
   */
  costPerHour: number;
  commission: number;
  travelMode: TravelMode;
  schedule: Array<ScheduleDate>;
}